requirements.txt
alembic==0.8.6
bcrypt==2.0.0
blinker==1.4
cffi==1.6.0
click==6.6
dateutils==0.6.6
decorator==4.0.10
dominate==2.2.0
Flask==0.10.1
Flask-Bootstrap==3.3.6.0
Flask-DebugToolbar==0.10.0
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==1.8.0
Flask-Principal==0.4.0
Flask-Script==2.0.5
Flask-Security==1.7.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
geohelper==0.2.0
gunicorn==19.6.0
infinity==1.4
intervals==0.7.1
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.4
MarkupSafe==0.23
parsedatetime==2.1
passlib==1.6.5
pycparser==2.14
pyparsing==2.1.4
python-dateutil==2.5.3
python-editor==1.0
pytz==2016.6.1
requests==2.10.0
six==1.10.0
SQLAlchemy==1.0.13
SQLAlchemy-Searchable==0.10.1
SQLAlchemy-Utils==0.32.7
texttable==0.8.4
Unidecode==0.4.19
validators==0.10.1
virtualenv==1.11.6
visitor==0.1.3
Werkzeug==0.11.10
WTForms==2.1
WTForms-Alchemy==0.16.1
WTForms-Components==0.10.0
mysqlclient==1.3.7
beautifulsoup4==4.4.1
flask-uploads
python-slugify
LatLon
git+http://dev.sigpipe.me/DashieHam/pyHamQth.git#egg=pyhamqth
git+https://github.com/ggramaize/libqth.git#egg=libqth
